requirements.txt
streamlit
numpy
pillow
opencv-python-headless
tensorflow
python-dotenv
pandas
requests
matplotlib
reportlab